+++ tests/data_model/wikibase/test_value_matchers.py
@@ -1,54 +1,8 @@
 import pytest
 
-from wikidatarefisland.data_model.value_matchers import ValueMatchers, TextValue, QuantityValue
+from wikidatarefisland.data_model.wikibase.value_matchers import ValueMatchers
 
-mock = {
-    "statement": {
-        "with_quantity": {
-            "datatype": 'quantity',
-            "value": {
-                "amount": "12"
-            }
-        },
-        "with_string": {
-            "datatype": 'string',
-            "value": {
-                "value": "Test"
-            }
-        },
-        "with_url": {
-            "datatype": 'url',
-            "value": {
-                "value": "Test"
-            }
-        },
-        "with_monolingualtext": {
-            "datatype": 'monolingualtext',
-            "value": {
-                "value": {
-                    "text": "Test"
-                }
-            }
-        },
-        "without_type": {
-            "datatype": 'some-other-data'
-        }
-    },
-    "reference": {
-        "with_one_quantity_match": {
-            "extractedData": ["12"]
-        },
-        "with_one_string_match": {
-            "extractedData": ["Test"]
-        },
-        "with_multiple_values_match": {
-            "extractedData": ["12", "Test"]
-        },
-        "without_match": {
-            "extractedData": ["Some", "Other", "Values"]
-        }
-    }
-}
+from mock_data import mock
 
 given = {
     "no_type": {
@@ -87,28 +41,6 @@ given = {
         "reference": mock["reference"]["with_multiple_values_match"]
     }
 }
-
-
-class TestQuantityValue:
-
-    @pytest.mark.parametrize("statement,equivalent", [
-        (mock["statement"]["with_quantity"], "12")
-    ])
-    def test_equivalence(self, statement, equivalent):
-        assert QuantityValue(statement) == equivalent
-
-
-class TestTextValue:
-
-    @pytest.mark.parametrize("statement,equivalent", [
-        (mock["statement"]["with_string"], "Test"),
-        (mock["statement"]["with_url"], "Test"),
-        (mock["statement"]["with_monolingualtext"], "Test"),
-        (mock["statement"]["with_string"], "test"),
-        (mock["statement"]["with_string"], "  Test ")
-    ])
-    def test_equivalence(self, statement, equivalent):
-        assert TextValue(statement) == equivalent
 
 
 class TestValueMatchers:
