+++ wikidatarefisland/data_model/__init__.py
@@ -1,7 +1,3 @@
-from .value_matchers import ValueMatchers, TextValue, QuantityValue
 from .schemaorg_normalizer import SchemaOrgNormalizer, SchemaOrgGraph, SchemaOrgNode
 
-__all__ = [
-    SchemaOrgNormalizer, SchemaOrgGraph, SchemaOrgNode,
-    ValueMatchers, TextValue, QuantityValue
-]
+__all__ = [SchemaOrgNormalizer, SchemaOrgGraph, SchemaOrgNode]
